+++ modules/ibkr/src/test/scala/com/odenzo/ibkr/web/flexquery/parsing/XMLParsingTest.scala
@@ -1,0 +1,8 @@
+package com.odenzo.ibkr.web.flexquery.parsing
+
+import com.odenzo.ibkr.web.flexquery.BaseTest
+
+class XMLParsingTest extends BaseTest {
+
+  test("Foo") {}
+}
